extract whisper file creation to a function
the extracted function can then be used when uncached metrics come in, to proactively create the whisper file

* move whisper file creation to a function in storage.py
* move loading storage and aggregation schemas to storage.py
